Add Strs helper for Go 1.6 cgo pointer passing rules.
This new helper creates C strings in the C heap using malloc, unlike
the normal Str helper that uses Go memory.

It's possible to pass it to cgo, for example, for gl.ShaderSource. This
is intended usage:

	csources, free := gl.Strs(source) // source string ends with "\x00"
	gl.ShaderSource(shader, 1, csources, nil)
	free()

Or for multiple parameters:

	csources, free := gl.Strs(sources...) // all strings end with "\x00"
	gl.ShaderSource(shader, len(sources), csources, nil)
	free()

It returns a free func that must be called once the string is no longer
needed to free memory.

Pair programmed implementation of Strs with @slimsag.

Fixes go-gl/gl#31.

Generated using changes in go-gl/glow#69.