Make the workflow a bit more reliable
This makes a few changes.
Most importantly, it fixes the installtion of deps.
This was due to the repo list not being updated and partly because apt-get
is a terrible package manager (in my opinion) and thus doesn't do it automatically.

The tests now only run for the oldest and newest Go versions only as if it works on both of those,
we can litterally say that anything between is a bug in Go and not here. Apart from that, it
is mostly some cleanups and updates to the used actions.